Move buf subsystem out of the lint subpackage
Hoists `subsystem.py` and `skip_field.py` from
`backend/codegen/protobuf/lint/buf/` up to
`backend/codegen/protobuf/buf/` so non-lint consumers (e.g. an upcoming
buf-based code generator) can depend on the Buf subsystem without
reaching into the lint package. Pure file relocation plus import path
updates; no behavior change.
