+++ src/python/pants/backend/codegen/protobuf/buf/BUILD
@@ -1,0 +1,4 @@
+# Copyright 2026 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+python_sources()
